requirements.txt
Django==4.2.26
djangorestframework==3.16.1
cryptography==46.0.3
